// src/core/field-engine/core/input-base/input-base.types.ts
export interface IInputProperties extends IFieldDescriptor {
    [key: string]: any
    id: number
    name: string
    label: string

    enabled: boolean
    isValid: boolean
    isDirty: boolean
    isPristine: boolean
    isFocus: boolean

    value: FieldDataTypes
    originalValue: FieldDataTypes

    /** message helper method : uses treacker and fallbacks to console */
    message: (type: TrackingType, source: string, message: string) => void
    /** Core methods */
    setValue: (value: FieldDataTypes) => void
    getValue: () => FieldDataTypes

    setFocus: () => void
    clear: () => void
    enable: (enabled: boolean) => void

    hasChanges: (callback: () => void) => void
    handleValidation: <T extends IEvents>(event?: T, data?: IValidationStrategyData) => void

    handleOnBlur: <T extends IEvents>(data?: T) => void
    handleOnFocus: <T extends IEvents>(data?: T) => void
    handleOnClear: <T extends IEvents>(data?: T) => void

    /** Dependency accessors */
    domManager: IDomManager<HTMLInputElement>
    drawer: IDrawerBaseInput
    styleManager: IStyleManager
    notificationManager: INotificationManager
    trackingManager: ITrackingManager
    validationManager: IValidationManager
    vlaueManager: IValueManager
}

export type SchemeToDescriptorConverterType = (scheme: IEntityScheme) => IFieldDescriptor

/** @warning: should not being used outside it's main implementation prefer the use of @Ifield */
export type IInput = IInputBase & Omit<IFieldDescriptor, 'validationOptions' | 'options'>

export interface IInputBase extends IInputProperties, IInitializableDependency {
    new (descriptor: IFieldDescriptor): IInputBase

    /** initializer builders */
    initializeProperties: (descriptor: IFieldDescriptor) => void
    checkInitialized: () => IInitilizationCheckResult
    useDomManager: (DomManagerInstance?: IDomManager<HTMLInputElement>) => IInputBase
    useNotificationManager: (notifierInstance?: INotificationManager) => IInputBase
    useTrackingManager: (trackerInstance?: ITrackingManager) => IInputBase
    useValidationManager: (validationStrategyInstance?: IValidationManager) => IInputBase
    useValueManager: (valueStrategyInstance?: IValueManager) => IInputBase
    useDrawerManager: (drawerableInstance?: IDrawerBaseInput) => IInputBase
    useStyleManager: (stylerInstance?: IStyleManager) => IInputBase
}

export interface IExtendedInputBase extends IInitializableDependency {
    input: IInputBase
}

export interface IExtendedInput
    extends IFieldStyleProperties,
        IOptionBaseInputProperties,
        IDrawerBaseInputProperties,
        IClickBaseInputProperties,
        ICheckBoxBaseInputProperties,
        IRadioBaseInputProperties,
        ISelectBaseInputProperties,
        IValueManagerProperties,
        ITextBaseInput,
        IExtendedInputBase,
        IInitializableDependency {
    handleOnClear: any
    // getOptionByValue: (value: string) => IOptionItem | null
    // tryGetOptionByIdOrValue: (id: string, value: string) => IOptionItem | null
    // handleOnChanged: <T extends IEvents>(data?: T) => void
    // handleOnClick: <T extends IEvents>(data?: T) => void
    // getOptionById: (id: string) => IOptionItem | null
}